+++ vstgui/lib/platform/win32/win32dragcontainer.cpp
@@ -179,7 +179,7 @@ uint32_t WinDragContainer::getData (uint32_t index, const void*& buffer, Type& t
 		}
 		buffer = strings[index].c_str ();
 		type = stringsAreFiles ? kFilePath : kText;
-		return (int32_t)strings[index].length ();
+		return static_cast<uint32_t> (strings[index].length ());
 	}
 	return 0;
 }
@@ -208,7 +208,7 @@ bool WinDragContainer::checkResolveLink (const TCHAR* nativePath, TCHAR* resolve
 				hres = ppf->Load (nativePath, STGM_READ);
 				if (SUCCEEDED (hres))
 				{					
-					hres = psl->Resolve (0, MAKELONG (SLR_ANY_MATCH | SLR_NO_UI, 500));
+					hres = psl->Resolve (0, static_cast<DWORD> (MAKELONG (SLR_ANY_MATCH | SLR_NO_UI, 500)));
 					if (SUCCEEDED (hres))
 					{
 						// Get the path to the link target.
